requirements.txt
# PyTorch с поддержкой CUDA 12.1 (совместимо с CUDA 13.0)
# Для установки используйте команду из https://pytorch.org/get-started/locally/
# Или установите напрямую:
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
pillow>=9.5.0
scikit-learn>=1.3.0
tqdm>=4.65.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0

# Для оптимизатора SPAM (Spike-Aware Adam with Momentum Reset)
# Установите дополнительно:
# pip install git+https://github.com/TianjinYellow/SPAM-Optimizer.git

